socketio_server/requirements.txt
#app
flask == 2.3.2
Flask-Cors == 4.0.0

#Redis
redis == 5.0.1

#socket.io
python-socketio == 5.10.0
